requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.3.0
attrs==23.2.0
black==24.3.0
certifi==2024.2.2
deepgram-sdk==3.2.3
distro==1.9.0
fastapi==0.110.0
httpcore==1.0.5
httpx==0.27.0
multidict==6.0.5
mypy-extensions==1.0.0
openai==1.14.3
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
requests==2.31.0
sniffio==1.3.1
starlette==0.36.3
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
python-multipart==0.0.9
Jinja2==3.1.3
python-dotenv==1.0.1